Added comments and some refactoring of tests to separate them into more clear modules
