+++ control_server/tests/client/init_test.py
@@ -1,4 +1,4 @@
-from control_server.tests.utils import get_prefix
+from control_server.tests.utils.generic_test_utils import get_prefix
 
 
 def test_init(client):
