+++ control_server/tests/database/mongo_database_test.py
@@ -1,7 +1,7 @@
 import pytest
 
-from control_server.tests.database_test_utils import DatabaseTestData
-from control_server.tests import database_test_utils
+from control_server.tests.database.database_test_utils import DatabaseTestData
+from control_server.tests.database import database_test_utils
 
 
 @pytest.fixture
@@ -10,7 +10,7 @@ def mongo_test_data():
     Set up testing data, and tear down afterwards
     :return: Nothing.
     """
-    data = DatabaseTestData(use_mock=True)
+    data = DatabaseTestData(use_mock=False)
     data.db.clear()
     yield data
     data.db.clear()
@@ -20,6 +20,9 @@ def test_set_delete_user(mongo_test_data: DatabaseTestData):
     """
     Tests that the set_user method works as expected without raising errors
     """
+    if mongo_test_data.settings.mock_db:
+        pytest.skip("Skipping test for mongodb database")
+
     database_test_utils.test_set_delete_user(test_data=mongo_test_data)
 
 
@@ -28,6 +31,9 @@ def test_set_get_all(mongo_test_data: DatabaseTestData):
     Tests that the set_user method works as expected without raising errors,
     and that a set user can be retrieved
     """
+    if mongo_test_data.settings.mock_db:
+        pytest.skip("Skipping test for mongodb database")
+
     database_test_utils.test_set_get_all(test_data=mongo_test_data)
 
 
@@ -36,4 +42,7 @@ def test_set_get_delete_get_user(mongo_test_data: DatabaseTestData):
     Tests that the set_user method works as expected without raising errors,
     and that a set user can be retrieved
     """
+    if mongo_test_data.settings.mock_db:
+        pytest.skip("Skipping test for mongodb database")
+
     database_test_utils.test_set_get_delete_get_user(test_data=mongo_test_data)
