+++ control_server/tests/utils/hashable_dict.py
@@ -1,4 +1,8 @@
 class HashableDict(dict):
+    """
+    Hashable wrapper for the dict class, allowing it to be used as a key in a
+    dictionary.
+    """
     def __hash__(self):
         return hash(tuple(sorted(self.items())))
